feat: Reorganise tests and resources for /ping
