+++ src/lambda/http/ping.ts
